adds virtual replay contexts to the GarlicMessageHandler
This patch adds unique msgIDXor values to the GarlicMessageHandler corresponding to the Local, Router, and Tunnel handler contexts.
This is related to the previous patch, wherein we added unique msgIDXor values to the messages we pass to the InNetMessagePool.add
function. For Garlic Messages, it is important to think about who we're getting them from and who we're sending them to, many of
them will be addressed to us for instance, and these cases are handled differently per delivery mode for now. This may not be necessary,
the known issue is that they are checked against the same pool as the exploratory tunnels. It looks like we might be better off to
give every 4-arg add() call it's own xor value.