Plugins: allow plugin updates to come from torrent sources.
This is a simple(I think), backward-compatible way of updating plugins using torrents as sources.
Plugin developers can keep the existing update URL.
If both `http://host.i2p/plugin.su3` and `http://host.i2p/plugin.su3.torrent` exist, then the updater will
download the torrent and treat the `http://host.i2p/plugin.su3` as a webseed.
When the download is finished, the plugin update proceeds as normal.

The main use-case for this is for very large plugins that potentially come with a lot of static data.
This is true of `i2p.plugins.firefox` in particular which is about ~130mb.